refactor: replace modal component with a hook
