+++ src/pages/event/[id]/index.tsx
@@ -1,19 +1,9 @@
-import {
-  Text,
-  Center,
-  Flex,
-  Button,
-  Box,
-  useDisclosure,
-} from '@chakra-ui/react';
+import { Text, Center, Flex, Button, Box } from '@chakra-ui/react';
 import { CheckIcon, PlusSquareIcon } from '@chakra-ui/icons';
 import {
-  useDeletePaymentMutation,
   useQueryEventNameQuery,
   useQueryPaymentsQuery,
 } from '../../../generated/graphql';
-import Modal from '../../../components/Modal';
-import { useState } from 'react';
 import { liffVar } from '../../../components/LiffProvider';
 import AddWarikankun from '../../../components/AddFriend';
 import useFriendship from '../../../hooks/useFriendship';
@@ -23,6 +13,8 @@ import SumPrice from '../../../components/SumPrice';
 import Payments from '../../../components/Payments';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faShareSquare } from '@fortawesome/free-solid-svg-icons';
+import useDeleteModal from '../../../hooks/useDeleteModal';
+
 const Event = () => {
   const router = useRouter();
   const { id } = router.query;
@@ -32,21 +24,9 @@ const Event = () => {
   const { data: paymentsData } = useQueryPaymentsQuery({
     variables: { eventId: id },
   });
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const [deleteTarget, setDeleteTarget] = useState<number | null>(null);
-  const [Mutation, { loading: isDeleting, error: deleteError }] =
-    useDeletePaymentMutation();
   const liff = liffVar();
   const { isFriend } = useFriendship();
-
-  const deletePayment = () => {
-    if (deleteTarget) {
-      Mutation({ variables: { paymentId: deleteTarget } }).then(() => {
-        onClose();
-        router.reload();
-      });
-    }
-  };
+  const { renderDeleteModal, openModal, setDeleteTarget } = useDeleteModal();
 
   const handleShareClick = () => {
     if (!liff || !eventNameData?.events[0]) return;
@@ -71,14 +51,7 @@ const Event = () => {
 
   return (
     <>
-      <Modal
-        isOpen={isOpen}
-        onClose={onClose}
-        onClick={deletePayment}
-        setDeleteTarget={setDeleteTarget}
-        loading={isDeleting}
-        error={deleteError}
-      />
+      {renderDeleteModal()}
       <EventName id={id} />
       <Flex justifyContent="space-evenly" mt="3">
         <Button colorScheme="teal" onClick={() => router.push(`${id}/add`)}>
@@ -99,7 +72,11 @@ const Event = () => {
         <SumPrice id={id} />
       </Box>
       <Box mb="5">
-        <Payments id={id} setDeleteTarget={setDeleteTarget} onOpen={onOpen} />
+        <Payments
+          id={id}
+          setDeleteTarget={setDeleteTarget}
+          onOpen={openModal}
+        />
       </Box>
       <Box mb="5">
         <Center>
